pyproject.toml
[tool.poetry]
name = "capablerobot_usbhub"
version = "0.2.4"
description = "Host side driver for the Capable Robot Programmable USB Hub"
authors = ["Chris Osterwood <osterwood@capablerobot.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/CapableRobot/CapableRobot_USBHub_Driver"
keywords = ["usb", "circuitpython", "hardware"]
include = ["examples", "LICENSE", "CODE_OF_CONDUCT.md", "50-capablerobot-usbhub.rules", "requirements.txt"]

[tool.poetry.dependencies]
python = "^3.6"
pyusb = "^1.0.2"
construct = "^2.9.51"
pyyaml = "^5.3"
click = "^7.0"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
usbhub = "capablerobot_usbhub.console:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
